Bug 1412359 - Filter to matching device on remove. r=gl
The local device removal path used by RDM had a bug in its `findIndex` call
which caused it to always return `true` for the first device.

Effectively this meant that each separate device removal button always removed
the first device!  This would lead to all sorts of user confusion and UI
divergence.

Here we clean this up by allowing the caller (RDM in this case) to specify via a
callback which device is intended for removal.

MozReview-Commit-ID: 22VwEDZAXOa

UltraBlame original commit: 973e547ce20c556c234201413716badcd41cb0b8

+++ devtools/client/shared/devices.js
@@ -60,14 +60,24 @@ async function loadLocalDevices() {
 
 
 
+
 async function addDevice(device, type = "phones") {
   await loadLocalDevices();
   let list = localDevices[type];
   if (!list) {
     list = localDevices[type] = [];
   }
+
+  
+  let exists = list.some(entry => entry.name == device.name);
+  if (exists) {
+    return false;
+  }
+
   list.push(Object.assign({}, device));
   await asyncStorage.setItem(LOCAL_DEVICES, JSON.stringify(localDevices));
+
+  return true;
 }
 
 
@@ -81,9 +91,8 @@ async function removeDevice(device, type = "phones") {
     return false;
   }
 
-  let index = list.findIndex(item => device);
-
-  if (index === -1) {
+  let index = list.findIndex(entry => entry.name == device.name);
+  if (index == -1) {
     return false;
   }
